axios and api fetching to front end
