+++ ReactApp/315rental/src/pages/css/Home.css
@@ -69,7 +69,6 @@ footer {
     margin: 5%;
     padding: 5%;
     box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2);
-
     display: grid;
     grid-template-columns: 1fr 1fr 1fr 1fr;
     grid-template-rows: 1.3fr 1.3fr 0.4fr;
